Address/pyproject.toml
[project]
name = "address"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ecdsa>=0.19.1",
    "pycryptodome>=3.22.0",
    "ruff>=0.11.2",
]
